WIP Removing old downloads page code
