+++ frontend/packages/portal-frontend/src/dataPage/hooks/useAllDataHandlers.ts
@@ -3,12 +3,9 @@ import { deleteQueryParams } from "@depmap/utils";
 import qs from "qs";
 import { useCallback } from "react";
 import { getReleaseGroupFromSelection } from "src/common/utilities/helper_functions";
-import { TypeGroupOption } from "../../download/components/CheckboxPanel";
 import { FileSearchOption } from "../../download/components/FileSearch";
-import {
-  findReleaseVersionGroupName,
-  findReleaseVersions,
-} from "../../download/utils";
+import { findReleaseVersionGroupName, findReleaseVersions } from "../utils";
+import { TypeGroupOption } from "../models/types";
 
 export default function useReleaseNameAndVersionSelectionHandlers(
   releaseData: Release[],
