--- portal-backend/depmap/static/css/downloads/all_downloads.scss
@@ -1,166 +1,0 @@
-@import "../shared/variables";
-@import "../shared/narrow_react_table";
-
-.download-table {
-  @include narrow-react-table;
-
-  .ReactTable .rt-th,
-  .ReactTable .rt-td {
-    cursor: pointer;
-    white-space: normal;
-    display: flex;
-    align-items: center;
-    font-size: 10pt;
-    word-break: break-word; // overflow-wrap: break-word works in an isolated instance but not here. This (as opposed to break-all) seems to achieve the issue of not breaking a word if possible, but breaking if ony choice
-  }
-
-  .horizontal-center-container {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-  }
-}
-
-.all-downloads {
-  .glyphicon-resize-full {
-    cursor: pointer;
-  }
-}
-
-.downloads-title-div-wrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.version-selector {
-  float: left;
-  margin-bottom: 10px;
-}
-
-.version-selector-label {
-  margin-right: 5px;
-}
-
-.select-file-set-label {
-  font-size: 16px;
-  font-weight: bold;
-}
-
-.file-set-selector {
-  float: left;
-  margin-right: 42px;
-  margin-bottom: 10px;
-}
-
-.card_subheading {
-  padding-top: 10px;
-}
-
-.card_title {
-  word-break: break-word;
-  margin-bottom: 4px;
-}
-
-.card_position {
-  position: fixed;
-  display: inline-block;
-  overflow-wrap: break-word;
-  word-break: break-word;
-  overflow-y: auto;
-  top: 22px;
-  bottom: 10px;
-  background-color: #ffffff;
-  margin: 100px 50px 10px 00px;
-  padding: 5px 20px 10px 5px;
-  border: 1px solid gray;
-  -webkit-transition: all 0.3s;
-  -moz-transition: all 0.3s;
-  -ms-transition: all 0.3s;
-  -o-transition: all 0.3s;
-  transition: all 0.3s;
-
-  @media screen and (max-width: $md_width) {
-    position: relative;
-    width: 100%;
-    max-width: unset;
-    margin: unset;
-    margin-block-start: 15px;
-    margin-block-end: 15px;
-    top: unset;
-    bottom: unset;
-  }
-}
-
-.view_more_less {
-  padding: 0;
-}
-
-.dl_card_smaller {
-  bottom: $footer_height;
-}
-
-.sticky {
-  position: -webkit-sticky;
-  position: sticky;
-  top: 0;
-}
-
-.text-margin {
-  margin: 0px 0px 6px 0px;
-}
-
-.checklist-padding {
-  padding-top: 20px;
-}
-
-.highlight-color {
-  color: $depmap_popout;
-}
-
-.checklist-fontsize {
-  font-size: 10pt;
-}
-
-.downloads-icon-modal-fontsize {
-  font-size: 14pt;
-  padding-left: 8px;
-}
-
-.card_logo {
-  height: 16px;
-}
-
-.selected-row {
-  // Note to Andrew: check with Josephine if things you write here don't appear, they might be overwritten by css provided by the React table
-  outline: 2px solid $depmap_popout;
-}
-
-.striped-row {
-  &:nth-child(odd) {
-    background-color: #f2f2f2;
-  }
-}
-
-.all-downloads {
-  .dropdown-menu {
-    height: 70vh;
-    overflow: scroll;
-  }
-  .tab-active {
-    border-left: 5px solid #9a5b95;
-  }
-  .tab-inactive {
-    border-left: 5px solid #b9b9b9;
-    color: #b9b9b9;
-  }
-  .overlay {
-    -webkit-filter: grayscale(100%);
-    filter: grayscale(100%);
-    background-color: rgba(255, 255, 255, 0.7);
-    position: absolute;
-    z-index: 2;
-    width: 100%;
-    height: 100%;
-    cursor: pointer;
-  }
-}
