--- portal-backend/depmap/templates/downloads/all_downloads.html
@@ -1,36 +1,0 @@
-{% extends "nav_footer.html" %}
-{% import "downloads/macros.html" as macros with context %}
-
-
-{% block page_title %}
-    DepMap Data Downloads
-{% endblock %}
-
-{% block css %}
-    <link rel="stylesheet" href="https://unpkg.com/react-table@6.8.0/react-table.css">
-    {% assets filters="libsass", output="public/css/all_downloads.css", "css/downloads/all_downloads.scss", depends='css/shared/*.scss' %}
-        <link rel="stylesheet" href="{{ ASSET_URL }}">
-    {% endassets %}
-    {% assets filters="libsass", output="public/css/headliners.css", "css/downloads/headliners.scss" %}
-    <link rel="stylesheet" href="{{ ASSET_URL }}">
-    {% endassets %}
-{% endblock %}
-
-{% block meta_description %}
-    The latest quarterly release and archive of available datasets
-{% endblock %}
-
-{% block content %}
-    <div id="react-all-downloads-root" data-react-component-loaded-for-selenium="false"></div>
-{% endblock %}
-
-{% block js %}
-   <script id="react-download-data" type="application/json">
-  {
-    "termsDefinitions": {{ terms_definitions|tojson|safe }},
-    "bulkDownloadCsvUrl": {{ bulk_files_csv_url|tojson|safe }},
-    "mode": "{{ mode }}"
-  }
-  </script>
-    <script src="{{ webpack_url('download.js') }}"></script>
-{% endblock %}
